removed second guideline file and renamed the original
