--- src/lib/Pages/Details/GuideLine1-2.svelte
@@ -1,205 +1,0 @@
-<script>
-    import { IconLibrary } from "$lib";
-</script>
-<article class="content-wrapper">
-    <section>
-        <h2>Op tijd gebasseerde media</h2>
-        <p>richtlijn 1.2</p>
-        <IconLibrary name ="eye" />
-        <progress id="file" value="32" max="100"> 32% </progress>
-    </section>
-    <section>
-        <details>
-            <summary>1.2.1 Louter-geluid & louter-videobeeld (vooraf opgenomen)<input type="checkbox"><IconLibrary name ="fine" /></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.2 Ondertitels voor doven en slechthorenden (vooraf opgenomen)<input type="checkbox"><IconLibrary name ="fine" /></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.3 Audiodescriptie of media-alternatief (vooraf opgenomen)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.4 Ondertitels voor doven en slechthorenden (live)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.5 Audiodescriptie (vooraf opgenomen)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.6 Gebarentaal (vooraf opgenomen)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.7 Verlengde audiodescriptie (vooraf opgenomen)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-        <details>
-            <summary>1.2.8 Media-alternatief (vooraf opgenomen)<input type="checkbox"></summary>
-            <ul>
-                <li>
-                    <svg></svg> <h3>impact op gebruiker</h3>
-                    <p></p>
-                </li>
-                <li>
-                    <svg></svg> <h3>hoe te testen</h3>
-                    <dl>
-                        <dt>1</dt><dd></dd>
-                        <dt>2</dt><dd></dd>
-                        <dt>3</dt><dd></dd>
-                        <dt>4</dt><dd></dd>
-                    </dl>
-                </li>
-            </ul>
-            <section>
-                <h3>Hoe kun je dit oplossen</h3>
-                <p></p>
-                <p></p>
-            </section>
-        </details>
-    </section>
-</article>
-<style>
-    article{
-        border-radius: var(--section-border-radius);
-        box-shadow: var(--box-shadow);
-        padding: var(--average-padding);
-        background-color: var(--color-background-section);
-    }
-</style>
