+++ src/lib/Pages/Details/GuideLines.svelte
