+++ src/lib/index.js
@@ -18,8 +18,7 @@ export { default as TypeChart } from "./Pages/Home/TypeChart.svelte";
 export { default as MonthlySwitchButtons } from "./Pages/Home/MonthlySwitchButtons.svelte";
 
 /*DETAILS*/
-export { default as GuideLine11 } from "./Pages/Details/GuideLine1-1.svelte";
-export { default as GuideLine12 } from "./Pages/Details/GuideLine1-2.svelte";
+export { default as GuideLine11 } from "./Pages/Details/GuideLines.svelte";
 export { default as Filter } from "./Pages/Details/Filter.svelte";
 export { default as SiteOverview } from "./Pages/Home/SiteOverview.svelte";
